% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{pums_recode}
\alias{pums_recode}
\title{Recode ACS' PUMS Datasets to Forsyth Futures' Labeling}
\usage{
pums_recode(data)
}
\arguments{
\item{data}{The PUMS dataset object}
}
\description{
This function recodes the ACS' PUMS variables to match Forsyth Futures' labeling standards
}
\details{
Warning: Make sure to make a column titled 'year' before using pums_recode()
}
\examples{
 year <- 2019
 tidycensus::get_pums(variables = c("RAC1P","HISP","SEX","POVPIP","AGEP","SCHL"), state = "NC", puma = c("01801","01802","01803"),
          year = year, survey = "acs1", rep_weights = "person") \%>\%
   dplyr::rename_all(~str_to_upper(.x)) \%>\%
   # Make sure to make a column titled 'year' before using pums_recode()
   dplyr::mutate(year = year) \%>\%
   pums_recode()
}
\keyword{pums_recode}
